Implements requirements.txt parser
